remove unnecessary RequestId expect/actual declarations
